+++ tests/testthat/test-reporter-progress.R
@@ -4,7 +4,7 @@ test_that("captures error before first test", {
   expect_known_output(
     test_file(
       test_path("reporters/error-setup.R"),
-      ProgressReporter$new(show_praise = FALSE)
+      ProgressReporter$new(show_praise = FALSE, min_time = Inf)
     ),
     test_path("reporters/progress-error-setup.txt")
   )
@@ -17,7 +17,7 @@ test_that("gracefully handles multiple contexts", {
   expect_known_output(
     test_file(
       test_path("reporters/context.R"),
-      ProgressReporter$new(show_praise = FALSE)
+      ProgressReporter$new(show_praise = FALSE, min_time = Inf)
     ),
     test_path("reporters/progress-context.txt")
   )
@@ -29,8 +29,22 @@ test_that("fails after max_fail tests", {
   expect_known_output(
     test_file(
       test_path("reporters/fail-many.R"),
-      ProgressReporter$new(show_praise = FALSE)
+      ProgressReporter$new(show_praise = FALSE, min_time = Inf)
     ),
     test_path("reporters/progress-fail-many.txt")
+  )
+})
+
+# compact display ---------------------------------------------------------
+
+test_that("fails after max_fail tests", {
+  local_reproducible_output(unicode = TRUE)
+
+  expect_known_output(
+    test_file(
+      test_path("reporters/tests.R"),
+      CompactProgressReporter$new()
+    ),
+    test_path("reporters/compact.txt")
   )
 })